+++ tycho-bnd-plugin/src/main/java/org/eclipse/tycho/bnd/mojos/BndIntegrationTestMojo.java
@@ -1,0 +1,34 @@
+/*******************************************************************************
+ * Copyright (c) 2023 Christoph Läubrich and others.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *    Christoph Läubrich - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.tycho.bnd.mojos;
+
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.ResolutionScope;
+
+import aQute.bnd.build.Project;
+import aQute.bnd.osgi.Constants;
+
+@Mojo(name = "integration-test", defaultPhase = LifecyclePhase.INTEGRATION_TEST, requiresDependencyResolution = ResolutionScope.TEST, threadSafe = true)
+public class BndIntegrationTestMojo extends AbstractBndProjectMojo {
+
+	@Override
+	protected void execute(Project project) throws Exception {
+		String testcases = project.getProperty(Constants.TESTCASES);
+		if (testcases == null) {
+			return;
+		}
+		project.test();
+	}
+
+}